handling get text error and rename drag command
