+++ test_driver/app.dart
@@ -1,4 +1,4 @@
-import 'package:demo/extended_commands.dart';
+import 'package:demo/drag_commands.dart';
 import 'package:flutter_driver/driver_extension.dart';
 import 'package:demo/main.dart' as app;
